requirements.txt
# SANZERO 경량화 의존성 (핵심 AI 기능만)

# FastAPI 코어
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.25.0
jinja2>=3.1.0,<3.2.0

# 인증 및 보안
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.0,<1.8.0
python-multipart>=0.0.5,<0.1.0

# 데이터베이스
supabase>=2.3.0,<2.4.0
gotrue>=2.8.0,<2.9.0

# HTTP 클라이언트
httpx>=0.24.0,<0.26.0

# 데이터 검증 및 처리
pydantic>=2.3.0,<2.4.0
pydantic-settings>=2.0.0,<2.1.0
email-validator>=2.0.0,<2.1.0

# 보안
bleach>=6.0.0,<7.0.0
cryptography>=40.0.0,<47.0.0

# 유틸리티 및 로깅
python-dotenv>=1.0.0,<1.1.0
loguru>=0.7.0,<0.8.0

# AI/LLM 연동
anthropic>=0.28.0,<1.0.0
openai>=1.10.0,<2.0.0

# HTTP 및 JSON 처리
requests>=2.31.0,<2.33.0

# 기본 ML 및 데이터 처리
joblib>=1.4.0,<2.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<1.7.0

# AI 모델 의존성 (CPU 버전만)
tensorflow-cpu>=2.16.1,<2.21.0
# keras는 TensorFlow 2.16+에서 내장됨
kmodes>=0.12.2,<0.13.0

# NLP 및 임베딩 (BERT 유사도 검색용) - v3 호환 버전 - 임시 주석
# sentence-transformers>=2.2.0,<3.0.0
# transformers>=4.20.0,<5.0.0
# torch>=2.0.0,<3.0.0

# 스케줄링 및 자동화
APScheduler>=3.10.0,<3.11.0

# 웹 크롤링 (핵심만)
beautifulsoup4>=4.12.0,<4.13.0
lxml>=4.9.0,<5.0.0

# 파일 처리
aiofiles>=23.2.0,<24.0.0

# PDF 레포트 생성
reportlab>=4.0.0,<4.3.0
matplotlib>=3.7.0,<3.9.0

# 테스트
pytest>=7.4.0,<7.5.0
pytest-asyncio>=0.21.0,<0.22.0